requirements.txt
aiohttp==3.7.2
aiohttp-cors==0.7.0
aiohttp-session==2.9.0
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
gino==1.0.1
motor==2.3.1
psycopg2-2.8.6
pymongo==3.11.3
requests==2.25.1
websocket==0.2.1
websocket-client==0.56.0
